Move code to commands mod
